% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impeachment-dilma-data.R
\docType{data}
\name{impeachmentDilma}
\alias{impeachmentDilma}
\title{Dilma's impeachment lower chamber approval data}
\format{An object of class \code{"data.frame"}.}
\source{
Brazilian lower chamber: \href{https://www.http://camara.gov.br/}{CD}
}
\usage{
data("impeachment-dilma")
}
\description{
Data from Brazilian lower chamber web service on President Dilma Rousseff
impeachment approvement on April 17h, 2016. The \code{data.frame} contains
the content and meta-data of all recorded speeches (not only the votes).
}
\examples{
data("impeachment-dilma")
}
\keyword{datasets}
